Backend/requirements.txt
Django>=5.1.2
djangorestframework
django-cors-headers
python-dotenv
cryptography
pymysql
pymongo
requests
PyJWT
num2words
pytz
sqlparse
asgiref
tzdata
Pillow
django-filter
pandas
openpyxl
twilio